+++ frontend/src/js/controllers/analyticsCtrl.js
@@ -247,7 +247,7 @@
         };
 
         vm.downloadChallengeParticipantTeams = function() {
-            parameters.url = "analytics/challenge/" + vm.challengeId + "/download_all_participants";
+            parameters.url = "analytics/challenge/" + vm.challengeId + "/download_all_participants/";
                 parameters.method = "GET";
                 parameters.callback = {
                     onSuccess: function(response) {
